server/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
openenv-core>=0.2.1
openai>=1.0.0
huggingface_hub>=0.20.0
pytest>=8.0.0
